Merged main into PlayerShooting.
Had to fix some code errors (left in the conflict stuff in WeaponController.cs) and one more conflict for projectiles.

Changed the enemies in the scene to face foward with the guns positioned correctly. Changed some values on the Enemy prefabs to be correct for enemy firing. Should be good to merge back into main now.